Increased Building Automation
Worked on improving the image-building process to be more automated.
Removed the need for a user to run an additional script once the
container was set up. Also added a non-root user to run the lab in.
Modified runtime of container to enter into a bash terminal instead of
invoking the Python interpreter.

The README was updated to reflect these changes and give instructions
based on the new building and running process.

Signed-off-by: wheeler-cs <wheelercs@protonmail.com>